Get release builds working by undefining NDEBUG
Shrinkler sources don't compile properly if NDEBUG is defined. This
is due to a very peculiar custom assert() macro in the Shrinkler sources.
CMake in turn insists in defining NDEBUG in release builds.

As a workaround we undefine NDEBUG before we include shrinkler sources.
Of course doing things like that is a recipe for getting ODR violations.